+++ src/main/java/com/apiround/greenhub/repository/RecipeRepository.java
@@ -4,9 +4,9 @@ import com.apiround.greenhub.entity.Recipe;   // ✅ Recipe 엔티티 import
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
     List<Recipe> findByStatusOrderByRecipeIdDesc(String status);
 
     // 마이페이지 - 내가 쓴 레시피
-    List<Recipe> findAllByUserUserId(Long userId);
+    List<Recipe> findAllByUserUserId(Integer userId);
 }
