+++ src/main/java/com/apiround/greenhub/repository/RecipeStepRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RecipeStepRepository extends JpaRepository<RecipeStep, Long> {
+public interface RecipeStepRepository extends JpaRepository<RecipeStep, Integer> {
 
     // 엔티티의 자바 필드명 기준 (recipe.recipeId, stepNo)
-    List<RecipeStep> findByRecipeRecipeIdOrderByStepNoAsc(Long recipeId);
+    List<RecipeStep> findByRecipeRecipeIdOrderByStepNoAsc(Integer recipeId);
 }
