+++ src/main/java/com/apiround/greenhub/repository/RecipeXProductRepository.java
@@ -4,6 +4,6 @@ import com.apiround.greenhub.entity.RecipeXProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface RecipeXProductRepository extends JpaRepository<RecipeXProduct, Long> {
-    List<RecipeXProduct> findByRecipeRecipeId(Long recipeId);
+public interface RecipeXProductRepository extends JpaRepository<RecipeXProduct, Integer> {
+    List<RecipeXProduct> findByRecipeRecipeId(Integer recipeId);
 }
